[AEFD specific]Added a parsing to identify all the ACT internal messages
The AEFD format is the format currently used in the company where this
sofwtare is developped. There is nothing interresting about it, it is
simply internal business.

- Added this function in GraphFactoryAEFD
- Added more documentation into GraphFactoryAEFD
- Created an implementation for Condition. It is the simplest thing
possible that simply stores the String in order to be able to compare
the file to be loaded and the actual loaded model.

TODO : there still need to order the transitions