Checking integrity when loading an AEFD file
When loading an AEFD file, we print out a new file from the extracted
data and we compare the new result with the initial one.

From now, we does not strictly write the model, but only the transitions
that should have been added into the model.
For a strict verification, we should also check that transitions added
to the model are correctly added.